+++ tests/context.py
@@ -4,4 +4,4 @@ import sys
 import os
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-import dpkg
+import biotracks
